requirements.txt
mss
Pillow
numpy
pygame
